added missing tz in test, moved OneDaysTrackerMotion for consistency
